+++ packages/core/src/config/config.service.ts
@@ -1,10 +1,10 @@
-import { DynamicModule, Injectable, Type } from '@nestjs/common';
-import { CorsOptions } from '@nestjs/common/interfaces/external/cors-options.interface';
-import { RequestHandler } from 'express';
+import {DynamicModule, Injectable, Type} from '@nestjs/common';
+import {CorsOptions} from '@nestjs/common/interfaces/external/cors-options.interface';
+import {RequestHandler} from 'express';
 
-import { getConfig } from './config-helpers';
-import { EntityIdStrategy } from './entity-id-strategy/entity-id-strategy';
-import { Logger, PickerLogger } from './logger/picker-logger';
+import {getConfig} from './config-helpers';
+import {EntityIdStrategy} from './entity-id-strategy/entity-id-strategy';
+import {Logger, PickerLogger} from './logger/picker-logger';
 import {
     ApiOptions,
     AssetOptions,
@@ -13,6 +13,8 @@ import {
 } from './picker-config';
 import {CustomFields} from "@picker-cc/common/lib/shared-types";
 import {GraphQLSchema} from "graphql";
+import {ModuleRef} from "@nestjs/core";
+import {Injector} from "../common";
 
 /**
  * 配置服务，一个工厂模式方便应用系统的全局配置
@@ -21,12 +23,14 @@ import {GraphQLSchema} from "graphql";
 export class ConfigService implements PickerConfig {
     private activeConfig: RuntimePickerConfig;
 
-    constructor() {
+    constructor(
+        private moduleRef: ModuleRef
+    ) {
         this.activeConfig = getConfig();
         // console.log('配置服务开启。。。。')
         // if (this.activeConfig.authOptions.disableAuth) {
-            // tslint:disable-next-line
-            // Logger.warn('Auth 已被禁用，在生产系统中不应该出现这种情况!');
+        // tslint:disable-next-line
+        // Logger.warn('Auth 已被禁用，在生产系统中不应该出现这种情况!');
         // }
     }
 
@@ -40,15 +44,25 @@ export class ConfigService implements PickerConfig {
     //         }, 200)
     //     })
     // }
+
+    get injector() {
+        const injector: any = new Injector(this.moduleRef);
+        return injector
+        // return this.activeConfig.injector
+    }
+
     get schemaConfig() {
         return this.activeConfig.schemaConfig
     }
+
     get context() {
         return this.activeConfig.context
     }
+
     get graphqlSchema(): GraphQLSchema {
         return this.activeConfig.graphqlSchema
     }
+
     get apiOptions(): Required<ApiOptions> {
         return this.activeConfig.apiOptions;
     }
@@ -56,6 +70,7 @@ export class ConfigService implements PickerConfig {
     get shouldDropDatabase(): Required<boolean> {
         return this.activeConfig.shouldDropDatabase
     }
+
     get port(): number {
         return this.activeConfig.apiOptions.port;
     }